UAR-417: Address code review comments
