+++ src/main/java/uk/gov/companieshouse/overseasentitiesapi/model/updatesubmission/changelist/changes/managingofficer/CorporateManagingOfficerChange.java
@@ -3,4 +3,9 @@ package uk.gov.companieshouse.overseasentitiesapi.model.updatesubmission.changel
 import uk.gov.companieshouse.overseasentitiesapi.model.updatesubmission.changelist.changes.managingofficer.officer.CorporateManagingOfficer;
 
 public class CorporateManagingOfficerChange extends ManagingOfficerChange<CorporateManagingOfficer> {
+    private static final String APPOINTMENT_TYPE = "Corporate Managing Officer";
+
+    public CorporateManagingOfficerChange() {
+        super.appointmentType = APPOINTMENT_TYPE;
+    }
 }
